add backup & restore name field
