+++ tests/apschedule/test_task_backup.py
@@ -8,8 +8,6 @@ from src.models import (
     BackupFilenames,
     Backups,
     ErrorBackups,
-    ErrorRestoredBackups,
-    RestoredBackups,
 )
 
 
@@ -23,7 +21,7 @@ def test_task_backup_volume(mocker, session):
     mocker.patch("src.apschedule.tasks.BACKUP_DIR", "/backup")
     from src.apschedule.tasks import task_create_backup
 
-    task_create_backup("test-volume", "job_id_1")
+    task_create_backup("test-volume", "job_id_1", "job_name_1")
 
     mock_backup_volume.assert_called_once_with(
         "test-volume",
@@ -39,6 +37,7 @@ def test_task_backup_volume(mocker, session):
     assert backup_db
     assert backup_db.backup_id == "job_id_1"
     assert backup_db.backup_filename == f"test-volume-{dt_now.isoformat()}.tar.gz"
+    assert backup_db.backup_name == "job_name_1"
     assert backup_db.backup_created == dt_now.isoformat()
     assert backup_db.backup_path == f"/backup/test-volume-{dt_now.isoformat()}.tar.gz"
     assert backup_db.volume_name == "test-volume"
@@ -69,7 +68,7 @@ def test_task_backup_volume_error(mocker, session):
     from src.apschedule.tasks import task_create_backup
 
     with pytest.raises(Exception, match="test error"):
-        task_create_backup("test-volume", "job_id_1")
+        task_create_backup("test-volume", "job_id_1", "job_name_1")
 
     mock_backup_volume.assert_called_once_with(
         "test-volume",
@@ -83,6 +82,7 @@ def test_task_backup_volume_error(mocker, session):
     assert backup_db
     assert backup_db.backup_id == "job_id_1"
     assert backup_db.error_message == "test error"
+    assert backup_db.backup_name == "job_name_1"
 
 
 @freeze_time(lambda: datetime.now(timezone.utc), tick=False)
@@ -99,7 +99,12 @@ def test_task_backup_volume_error_schedule(mocker, session):
     from src.apschedule.tasks import task_create_backup
 
     with pytest.raises(Exception, match="test error"):
-        task_create_backup("test-volume", "job_id_1", "job_name_1", True)
+        task_create_backup(
+            "test-volume",
+            "job_id_1",
+            "job_name_1",
+            is_schedule=True,
+        )
 
     mock_backup_volume.assert_called_once_with(
         "test-volume",
@@ -113,6 +118,7 @@ def test_task_backup_volume_error_schedule(mocker, session):
     assert backup_db
     assert backup_db.backup_id == "test-uuid"
     assert backup_db.error_message == "test error"
+    assert backup_db.backup_name == "job_name_1"
 
 
 @freeze_time(lambda: datetime.now(timezone.utc), tick=False)
@@ -146,6 +152,7 @@ def test_task_backup_volume_schedule(mocker, session):
     assert backup_db.backup_path == f"/backup/test-volume-{dt_now.isoformat()}.tar.gz"
     assert backup_db.volume_name == "test-volume"
     assert backup_db.schedule_id == "job_id_1"
+    assert backup_db.backup_name == "job_name_1"
 
     db_backup_filenames = session.exec(
         select(BackupFilenames).where(BackupFilenames.backup_id == "test-uuid")
@@ -157,79 +164,3 @@ def test_task_backup_volume_schedule(mocker, session):
         == f"test-volume-{dt_now.isoformat()}.tar.gz"
     )
     assert db_backup_filenames.backup_id == "test-uuid"
-
-
-@freeze_time(lambda: datetime.now(timezone.utc), tick=False)
-def test_task_restore_backup(mocker, session):
-    mocker.patch(
-        "src.apschedule.tasks.Session",
-        **{"return_value.__enter__.return_value": session},
-    )
-    mock_restore_volume = mocker.patch("src.apschedule.tasks.restore_volume")
-    mocker.patch("src.apschedule.tasks.BACKUP_DIR", "/backup")
-    backup = Backups(
-        backup_id="job_id_1",
-        backup_filename="test-volume-2021-01-01T00:00:00.000000+00:00.tar.gz",
-        backup_created="2021-01-01T00:00:00.000000+00:00",
-        backup_path="/backup/test-volume-2021-01-01T00:00:00.000000+00:00.tar.gz",
-        volume_name="test-volume",
-    )
-    session.add(backup)
-
-    from src.apschedule.tasks import task_restore_backup
-
-    task_restore_backup(backup.volume_name, backup.backup_filename, "job_id_2")
-
-    mock_restore_volume.assert_called_once_with(
-        backup.volume_name, "/backup", backup.backup_filename
-    )
-    restore_db = session.exec(
-        select(RestoredBackups).where(RestoredBackups.restore_id == "job_id_2")
-    ).first()
-
-    dt_now = datetime.now(timezone.utc)
-
-    assert restore_db
-    assert restore_db.restore_id == "job_id_2"
-    assert restore_db.backup_filename == backup.backup_filename
-    assert restore_db.restored_date == dt_now.isoformat()
-    assert restore_db.volume_name == "test-volume"
-
-
-@freeze_time(lambda: datetime.now(timezone.utc), tick=False)
-def test_task_restore_backup_error(mocker, session):
-    mocker.patch(
-        "src.apschedule.tasks.Session",
-        **{"return_value.__enter__.return_value": session},
-    )
-    mock_restore_volume = mocker.patch(
-        "src.apschedule.tasks.restore_volume", side_effect=Exception("test error")
-    )
-    mocker.patch("src.apschedule.tasks.BACKUP_DIR", "/backup")
-    backup = Backups(
-        backup_id="job_id_1",
-        backup_filename="test-volume-2021-01-01T00:00:00.000000+00:00.tar.gz",
-        backup_created="2021-01-01T00:00:00.000000+00:00",
-        backup_path="/backup/test-volume-2021-01-01T00:00:00.000000+00:00.tar.gz",
-        volume_name="test-volume",
-    )
-    session.add(backup)
-
-    from src.apschedule.tasks import task_restore_backup
-
-    with pytest.raises(Exception, match="test error"):
-        task_restore_backup(backup.volume_name, backup.backup_filename, "job_id_2")
-
-    mock_restore_volume.assert_called_once_with(
-        backup.volume_name, "/backup", backup.backup_filename
-    )
-
-    restore_db = session.exec(
-        select(ErrorRestoredBackups).where(
-            ErrorRestoredBackups.restore_id == "job_id_2"
-        )
-    ).first()
-
-    assert restore_db
-    assert restore_db.restore_id == "job_id_2"
-    assert restore_db.error_message == "test error"
